# Reduce algorithm decomposer set lookups