@schlosna 2022-03-31
I'm not sure what is up with the OCA checks as I had previously signed circa 2011 as noted on https://oca.opensource.oracle.com/?ojr=old-contrib-list ; however, when https://oca.opensource.oracle.com/?ojr=contrib-list was created it looks like there was some manual transcription issue where `schlosna` incorrectly became `sehlosna`