@schlosna 2022-04-22
I emailed registrar@openjdk.java.net  on 2022-04-13 to try and fix the OCA  to GitHub user ID mapping, but haven't heard back yet.