@schlosna 2022-05-02
Rebased and ran the new `SSLHandshake` microbenchmarks and there's slight improvement in TLS resume on my local setup:

Before:
```
Benchmark                 (resume)  (tlsVersion)   Mode  Cnt     Score    Error  Units
SSLHandshake.doHandshake      true       TLSv1.2  thrpt   15  5472.225 ± 40.802  ops/s
SSLHandshake.doHandshake      true           TLS  thrpt   15   460.960 ±  9.083  ops/s
SSLHandshake.doHandshake     false       TLSv1.2  thrpt   15   289.168 ±  3.984  ops/s
SSLHandshake.doHandshake     false           TLS  thrpt   15   229.550 ±  6.251  ops/s
```

After:
```
Benchmark                 (resume)  (tlsVersion)   Mode  Cnt     Score    Error  Units
SSLHandshake.doHandshake      true       TLSv1.2  thrpt   15  5911.220 ± 79.167  ops/s
SSLHandshake.doHandshake      true           TLS  thrpt   15   463.598 ± 10.078  ops/s
SSLHandshake.doHandshake     false       TLSv1.2  thrpt   15   289.606 ±  2.508  ops/s
SSLHandshake.doHandshake     false           TLS  thrpt   15   225.396 ±  7.222  ops/s
```